Move HIV disease to HIV status and fix bugs
