+++ src/hivpy/column_names.py
@@ -66,8 +66,6 @@ MAX_CD4 = "max_cd4"                             # float: Maximum CD4 count to wh
 ART_NAIVE = "art_naive"                         # bool: True if person has never been on antiretroviral therapy
 X4_VIRUS = "x4_virus"                           # bool: True if X4 virus is present in person, false otherwise.
 
-
-
 DATE_OF_DEATH = "date_of_death"                 # None | datetime.date: date of death if dead, o/w None
 WHO3_EVENT = "who3_event"                       # Bool: True if who3 disease occurs this timestep in HIV positive person
 NON_TB_WHO3 = "non_tb_who3"                     # Bool: True if non-tb who3 disease occurs this timestep in HIV positive person
